requirements.txt
Django==1.9.6
django-bootstrap-form==3.2.1
django-contrib-comments==1.7.2
django-filter==0.14.0
django-model-utils==2.5.2
django-notifications-hq==1.2
django-pagination==1.0.7
django-postman==3.3.1
djangorestframework==3.4.5
django-threadedcomments==1.0.1
django-widget-tweaks==1.4.1
greenlet==0.4.10
jsonfield==1.0.3
lifter==0.2.1
psycopg2==2.6.2
pytz==2016.6.1
sphinx-me==0.3
